General Cleanup - Moved api code out of __init__ and into a separate script - Fix up calls related to the above - Fix related tests - Switched to gunicorn in the docker deploy
