+++ imagegw/entrypoint.sh
@@ -11,7 +11,7 @@ fi
 for service in $@ ; do
   echo "service: $service"
   if [ "$service"  == "api" ] ; then
-    python ./imagegwapi.py  &
+    gunicorn -b 0.0.0.0:5000 --backlog 2048 shifter_imagegw.api.api:app
   elif  [ $(echo $service|grep -c "worker:") -gt 0 ] ; then
     queue=$(echo $service|sed 's/.*://')
     echo "Worker Queue: $queue"
